+++ src/number_list.rs
@@ -1,15 +1,4 @@
-use crate::{Error, Result, Stream, WriteBuffer, WriteOptions};
-
-/// Representation of the [`<list-of-numbers>`] type.
-///
-/// [`<list-of-numbers>`]: https://www.w3.org/TR/SVG2/types.html#InterfaceSVGNumberList
-#[derive(Clone, PartialEq, Default)]
-pub struct NumberList(pub Vec<f64>);
-
-impl_from_vec!(NumberList, NumberList, f64);
-impl_vec_defer!(NumberList, f64);
-impl_display!(NumberList);
-impl_debug_from_display!(NumberList);
+use crate::{Result, Stream};
 
 /// A pull-based [`<list-of-numbers>`] parser.
 ///
@@ -50,54 +39,5 @@ impl<'a> Iterator for NumberListParser<'a> {
 
             Some(v)
         }
-    }
-}
-
-impl std::str::FromStr for NumberList {
-    type Err = Error;
-
-    fn from_str(text: &str) -> Result<Self> {
-        let mut vec = Vec::new();
-        for number in NumberListParser::from(text) {
-            vec.push(number?);
-        }
-
-        Ok(NumberList(vec))
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::ListSeparator;
-
-    #[test]
-    fn write_1() {
-        let list = NumberList(vec![1.0, 2.0, 3.0]);
-
-        let mut opt = WriteOptions::default();
-        opt.list_separator = ListSeparator::Space;
-
-        assert_eq!(list.with_write_opt(&opt).to_string(), "1 2 3");
-    }
-
-    #[test]
-    fn write_2() {
-        let list = NumberList(vec![1.0, 2.0, 3.0]);
-
-        let mut opt = WriteOptions::default();
-        opt.list_separator = ListSeparator::Comma;
-
-        assert_eq!(list.with_write_opt(&opt).to_string(), "1,2,3");
-    }
-
-    #[test]
-    fn write_3() {
-        let list = NumberList(vec![1.0, 2.0, 3.0]);
-
-        let mut opt = WriteOptions::default();
-        opt.list_separator = ListSeparator::CommaSpace;
-
-        assert_eq!(list.with_write_opt(&opt).to_string(), "1, 2, 3");
     }
 }